requirements.txt
# Procesamiento de datos
pandas
numpy
xarray
scipy
scikit-learn

# Datos geográficos
geopandas
shapely

# APIs y procesamiento web
aiohttp
requests
copernicusmarine

# Visualización
matplotlib
seaborn

# Manejo de archivos
netCDF4
h5netcdf
dask

# Desarrollo y testing
pytest
black
flake8
mypy

# Utilidades
python-multipart
python-dotenv
tqdm
fastapi
uvicorn
httpx
pytest
